rename multiprocess modules to be private
